requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.2
Django==4.1.7
django-rest-knox==4.2.0
djangorestframework==3.14.0
Pillow==9.4.0
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
dnspython==1.16.0
enum-compat==0.0.3
eventlet==0.25.2
greenlet==0.4.16
monotonic==1.5
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2020.1
six==1.15.0